Add files via upload
Invalid video URLs are now handled (albeit spaghettily). Also, videos that are much shorter than the bits cheered with the video (i.e. if someone cheers 1000 bits, good for 1000 seconds, with a 10 second video, it'll stop when the video is over rather than waiting the full 1000 seconds) are handled correctly.